InputContainer 구현 (GREEN)
App 컴포넌트 분리하여 Input Container와 List Container로 재구성함
